pyproject.toml
[project]
name = "live-stream-visualization"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "csp>=0.11.3",
    "dash>=3.2.0",
    "dash-bootstrap-components>=2.0.4",
    "databento>=0.63.0",
    "flask>=3.1.2",
    "flask-socketio>=5.5.1",
    "gunicorn>=20.1.0",
    "numpy>=1.26.4",
    "obspy>=1.4.2",
    "pandas>=2.3.2",
    "perspective-python[tornado]>=3.8.0",
    "plotly>=6.3.0",
    "polygon-api-client>=1.15.3",
    "pyarrow>=19.0.1",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "python-rapidjson>=1.21",
    "requests>=2.32.5",
    "streamlit>=1.49.1",
    "tornado>=6.5.2",
    "websocket-client>=1.8.0",
    "websockets>=14.2",
]
